Created 2 new files called Controller and Model and moved 2 files into Controller.
The 2 files were Program.cs and Game1.cs. Game1 was renamed to "SpaceGame" and then refactored in the file to match as well as changing the namespace to include "SpaceGame.Controller" due to the new layout of folders and naming.

Running the program right now will cause a blue window to appear.