Created a new class in the view folder called ParallaxingBackground.cs and changed/added things in SpaceGame.cs to make it work.
The ParallaxingBackground.cs file will make it so that there will be a background that is able to move around. The changed and additions to SpaceGame.cs makes it so that it is possible for the background to be created and run in the game window.

Next up is Adding Enemies