Added some methods and calls to make displays that show scores and health.
The program will calculate and draw the score and health and constantly update it based on what is going on in the game.
NOTE: Font has not been added to the game or implemented due to issues going on. Because of this the text display is well... not being displayed.

Next is Menus